Fix issue  #4 refactor docs
I added a few files and arranged the directory structure to be more in line
with the way readthedocs processes files.
I think that if you scan "files changes" you'll get a sense of how it all flows.
If you want to compare, you can pull this branch into your repo and tell readthedocs
to build this branch. It will treat it as an additional version and you can flip between the two
online.